BZR-93 implemented REST endpoints for authorization admin API

+++ src/main/kotlin/org/bazar/authorization/database/entity/ActionAttributeEntity.kt
@@ -4,7 +4,7 @@ import java.time.Instant
 
 data class ActionAttributeEntity(
     val id: Int,
-    val actionId: Long,
+    val actionId: Int,
     val name: String,
     val displayName: String,
     val valueType: String,
